add frontend skeleton
