+++ client/app/auth/login.tsx
@@ -1,0 +1,7 @@
+export default function Login() {
+  return (
+    <div>
+      <h1>Log in</h1>
+    </div>
+  );
+}
