+++ client/app/check-in/add-user.tsx
@@ -1,0 +1,7 @@
+export default function AddUser() {
+  return (
+    <div>
+      <h1>Add user</h1>
+    </div>
+  );
+}
