+++ client/app/check-in/check-in.tsx
@@ -1,0 +1,7 @@
+export default function CheckIn() {
+  return (
+    <div>
+      <h1>Check in</h1>
+    </div>
+  );
+}
