+++ client/app/lounge/activity.tsx
@@ -1,0 +1,7 @@
+export default function Activity() {
+  return (
+    <div>
+      <h1>Activity</h1>
+    </div>
+  );
+}
