+++ client/app/lounge/map.tsx
@@ -1,0 +1,7 @@
+export default function Map() {
+  return (
+    <div>
+      <h1>Map</h1>
+    </div>
+  );
+}
