+++ client/app/lounge/pc-info.tsx
@@ -1,0 +1,7 @@
+export default function PcInfo() {
+  return (
+    <div>
+      <h1>PC Info</h1>
+    </div>
+  );
+}
