+++ client/app/records/records.tsx
@@ -1,0 +1,7 @@
+export default function Records() {
+  return (
+    <div>
+      <h1>Records</h1>
+    </div>
+  );
+}
