+++ client/app/student-info/student-info.tsx
@@ -1,0 +1,7 @@
+export default function StudentInfo() {
+  return (
+    <div>
+      <h1>Student info</h1>
+    </div>
+  );
+}
